+++ src/main/kotlin/dev/adirelle/adicrafter/AdiCrafter.kt
@@ -5,6 +5,8 @@ import dev.adirelle.adicrafter.utils.Mod
 
 object AdiCrafter : Mod("adicrafter") {
 
+    val config = Config.loadOrCreate()
+
     override fun onInitialize() {
         feature(CrafterFeature)
     }