+++ src/main/kotlin/dev/adirelle/adicrafter/utils/Tickable.kt
@@ -1,0 +1,8 @@
+package dev.adirelle.adicrafter.utils
+
+import net.minecraft.server.world.ServerWorld
+
+interface Tickable {
+
+    fun tick(world: ServerWorld)
+}